requirements.txt
langchain-neo4j
langchain-google-genai
langchain-postgres
langchain-experimental
langchain-openai
langchain-chroma
langchain-huggingface
chromadb
sentence-transformers
transformers
fastapi
uvicorn
python-multipart
python-dotenv
streamlit
pymupdf
tiktoken
numpy
requests
tqdm
neo4j>=5.0.0
psycopg2-binary
google-generativeai